refactor
Suggestion:
Creating a new EpicsSignalRO every time `instrument_in_use()` is called seems too expensive. Make this a known signal but prefix it with "_" to indicate it is not for general use.  Also, since line was so long, switched that *one line* to common indentation style such as [black](https://pypi.org/project/black/).